Use artifacts to pass env to jobs
That way we can do everything important in the setup job and just
download the artifact in each subsequent job.

We now also just cache the wheels instead of the virtualenv itself; that
way we are guaranteed to have a fresh indico copy from the main repo
instead of whatever got cached some time ago.