VC/Zoom: Do not launch celery task after failure
When the request fails for some reason the RH does a rollback. In that
case the log entry is not in the DB, and thus cannot be passed to
Celery, which causes an error. On top of that, triggering the task is
not desirable in this case, because the event times haven't actually
been changed.