VC/Zoom: Get meeting/webinar type from Zoom (#256)
When people convert it directly on Zoom, Indico may not be aware of
this, and thus use the wrong APIs which do not work.

Now we always check the state on Zoon, and use that. There's a minor
overhead for webinars since we query the meeting API first, but for the
vast majority of events that use normal meetings this won't make a
difference.