Fix bad prometheus/livesync plugin interaction
When using the prometheus plugin but not livesync, the import of the
livesync plugin resulted in the task being registered w/ celery, and
thus resulted in the task executing and usually failing due to
non-existent tables.

Also, the prometheus plugin only checked if the livesync plugin was
importable, so even with it disabled it tried to query stats from it,
failing in case its tables did not exist.

Now livesync task initialization only happens at plugin init time, and
the prometheus plugin also only registers its metrics at init time so
the livesync metric is completely skipped if that plugin is disabled.